Refactor game screen
